# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Quiz Platform API",
        "description": "API documentation for Quiz Platform mobile app",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/dashboard": {
            "get": {
                "tags": [
                    "Admin Dashboard"
                ],
                "description": "Get admin dashboard statistics.",
                "operationId": "f6e3ffeb3c9fb988ca2a77ad48e2c04c",
                "responses": {
                    "200": {
                        "description": "Dashboard stats"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden - Admin only"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "description": "Enter your bearer token",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}